+++ packages/shared/src/react/hooks/index.ts
@@ -11,6 +11,6 @@ export { useReverification } from './useReverification';
 export { useStatements as __experimental_useStatements } from './useStatements';
 export { usePaymentAttempts as __experimental_usePaymentAttempts } from './usePaymentAttempts';
 export { usePaymentMethods as __experimental_usePaymentMethods } from './usePaymentMethods';
-export { useSubscriptionItems as __experimental_useSubscriptionItems } from './useSubscriptionItems';
 export { usePlans as __experimental_usePlans } from './usePlans';
+export { useSubscription as __experimental_useSubscription } from './useSubscription';
 export { useCheckout as __experimental_useCheckout } from './useCheckout';